Change response to NODATA when querying invalid IP
Previously, we would return FORMERR to let the user know they've typo'd the IP address that they want as NS record. However, services like Letsencrypt will query for e.g. `CAA 168.0.1.anyz.one` when you request a cert for `192.168.0.1.anyz.one`, which would lead to this format error, which is treated as a resolving failure, which leads to it refusing your certificate.

I guess we can also explain it as "this is not a valid IP, so there is nothing here", so returning a nodata response isn't *too* weird